Implement new `dist` module (#12)
This update includes a new `dist` module for computing the Hamming distance between two genotypes (aggregated over all loci). The initial implementation of the distance calculation is naive and unsophisticated: if the genotype call for an individual isn't exactly the same as the genotype call for a different individual at that locus, the distance is incremented by 1. We should consider improvements that take into account: 1) some alleles are more similar than other alleles; and 2) how to handle comparison of a homozygous allele vs a heterzygous allele (i.e. what is the distance between the homozygous genotype call `A,T,T,C` and the heterozygous call `A,T,T,C` and `G,T,T,C`).

Despite the naive implementation, preliminary testing suggests that it this approach is sufficient to distinguish matching genotypes from non-matches.

Partially fulfills #5.