Refactor genotype-related code (#21)
This update splits out the `Genotype` classes from the `sim` and `type` modules and creates a new dedicated module with two new classes: `SimulatedGenotype` and `ObservedGenotype`. The primary method of the `type` module was renamed from `genotype` to `type`, so that it is consistent with other modules (collisions with builtin functions notwithstanding). Closes #15.