Update usa8k simulation code (#48)
This update fixes and optimizes the usa8k simulation code to account for the latest changes to MicroHapulator. A few minimal scripts were added for batch processing the simulations, rather than making thousands of calls to the MicroHapulator CLI and incurring the extra overhead of initializing the Python interpreter that many times.