Fix USA panel (#56)
Previously the `usa` panel included 100 loci for which there was *supposed* to be allele frequency data for all 19 sub-populations in a mock population roughly matching demographics in the United States. Due to a bug in that code, some loci in the panel do not have allele frequency data available for all populations. Also, since the initial panel slated for evaluation on real data (`beta`) contains 50 loci, it makes sense to restrict the `usa` panel to 50 loci as well.

This update fixes the bug and limits the panel to the top 50 loci ranked by A<sub>e</sub>.

Related to #55.