Fix contrib API (#70)
While updating a demo of the Python API, I discovered that the `microhapulator.contrib.contrib` has no way of processing a `Profile` object already in memory. It could only take a JSON file or BAM/FASTA as input. This update fixes the API so that the `microhapulator.contrib.contrib` works *only* on `Profile` objects, and changes the CLI to accommodate all the requisite handling of input files.