Fix mhpl8r diff command (#75)
The `profile.alleles()` function returns `None` when a profile has no alleles for a marker. The `diff` module applies set operations on the output of this function without any error checking, which is problematic when a `None` is returned. This update modifies the function to return an empty set instead of `None`, so that set operations will work as expected.

The `mhpl8r diff` command was also updated so that each allele is printed on a dedicated line, rather than joined on the same line.

This update incidentally includes a few small changes to the location of certain package files.